added dpdk as a submodule

+++ build.sh
@@ -3,16 +3,16 @@
 # TODO: this should probably be a makefile
 # TODO: install target
 (
-# Compile dpdk
-if [ ! -d "./dpdk" ]; then
-echo "./dpdk doesn't exist"
-exit 1
+cd deps/dpdk
+if [[ ! -e Makefile ]]
+then
+	echo "ERROR: dpdk submodule not initialized"
+	echo "Please run git submodule update --init"
+	exit 1
 fi
-cd ./dpdk
 make -j 8 install T=x86_64-native-linuxapp-gcc
-
-# Compile ovs
-cd ../setup-scripts
+cd ../../
+cd setup-scripts
 source ./helpers/setup-vars-ovs-dpdk.sh
 cd ../
 ./boot.sh
